+++ src/main/java/com/nexters/goalpanzi/application/auth/dto/request/AppleLoginCommand.java
@@ -1,4 +1,4 @@
-package com.nexters.goalpanzi.application.auth.dto;
+package com.nexters.goalpanzi.application.auth.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotEmpty;
