+++ src/main/java/com/nexters/goalpanzi/application/auth/dto/request/RefreshTokenCommand.java
@@ -1,11 +1,9 @@
-package com.nexters.goalpanzi.application.auth.dto;
+package com.nexters.goalpanzi.application.auth.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotEmpty;
 
-public record TokenResponse(
-        @Schema(description = "access token", requiredMode = Schema.RequiredMode.REQUIRED)
-        @NotEmpty String accessToken,
+public record RefreshTokenCommand(
         @Schema(description = "refresh token", requiredMode = Schema.RequiredMode.REQUIRED)
         @NotEmpty String refreshToken
 ) {
