+++ src/main/java/com/nexters/goalpanzi/application/auth/dto/response/TokenResponse.java
@@ -1,9 +1,11 @@
-package com.nexters.goalpanzi.application.auth.dto;
+package com.nexters.goalpanzi.application.auth.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotEmpty;
 
-public record RefreshTokenCommand(
+public record TokenResponse(
+        @Schema(description = "access token", requiredMode = Schema.RequiredMode.REQUIRED)
+        @NotEmpty String accessToken,
         @Schema(description = "refresh token", requiredMode = Schema.RequiredMode.REQUIRED)
         @NotEmpty String refreshToken
 ) {
