+++ src/main/java/com/nexters/goalpanzi/application/mission/MissionVerificationService.java
@@ -1,13 +1,10 @@
 package com.nexters.goalpanzi.application.mission;
 
-import com.nexters.goalpanzi.application.member.dto.ProfileResponse;
-import com.nexters.goalpanzi.application.mission.dto.CreateMissionVerificationCommand;
-import com.nexters.goalpanzi.application.mission.dto.MissionDetailResponse;
-import com.nexters.goalpanzi.application.mission.dto.MissionResponse;
-import com.nexters.goalpanzi.application.mission.dto.MissionVerificationCommand;
-import com.nexters.goalpanzi.application.mission.dto.MissionVerificationResponse;
-import com.nexters.goalpanzi.application.mission.dto.MissionsResponse;
-import com.nexters.goalpanzi.application.mission.dto.MyMissionVerificationCommand;
+import com.nexters.goalpanzi.application.mission.dto.request.CreateMissionVerificationCommand;
+import com.nexters.goalpanzi.application.mission.dto.request.MissionVerificationCommand;
+import com.nexters.goalpanzi.application.mission.dto.response.MissionVerificationResponse;
+import com.nexters.goalpanzi.application.mission.dto.response.MissionsResponse;
+import com.nexters.goalpanzi.application.mission.dto.request.MyMissionVerificationCommand;
 import com.nexters.goalpanzi.domain.member.Member;
 import com.nexters.goalpanzi.domain.member.repository.MemberRepository;
 import com.nexters.goalpanzi.domain.mission.Mission;
