+++ src/main/java/com/nexters/goalpanzi/application/mission/dto/request/MyMissionVerificationCommand.java
@@ -1,4 +1,4 @@
-package com.nexters.goalpanzi.application.mission.dto;
+package com.nexters.goalpanzi.application.mission.dto.request;
 
 public record MyMissionVerificationCommand(
         Long memberId,
