+++ src/main/java/com/nexters/goalpanzi/application/mission/dto/response/MissionVerificationResponse.java
@@ -1,4 +1,4 @@
-package com.nexters.goalpanzi.application.mission.dto;
+package com.nexters.goalpanzi.application.mission.dto.response;
 
 import com.nexters.goalpanzi.domain.member.Member;
 import com.nexters.goalpanzi.domain.mission.MissionVerification;
