+++ src/main/java/com/nexters/goalpanzi/application/mission/dto/response/MissionsResponse.java
@@ -1,6 +1,6 @@
-package com.nexters.goalpanzi.application.mission.dto;
+package com.nexters.goalpanzi.application.mission.dto.response;
 
-import com.nexters.goalpanzi.application.member.dto.ProfileResponse;
+import com.nexters.goalpanzi.application.member.dto.response.ProfileResponse;
 import com.nexters.goalpanzi.domain.member.Member;
 import com.nexters.goalpanzi.domain.mission.MissionVerification;
 import io.swagger.v3.oas.annotations.media.Schema;
