+++ src/main/java/com/nexters/goalpanzi/presentation/auth/AuthControllerDocs.java
@@ -1,6 +1,10 @@
 package com.nexters.goalpanzi.presentation.auth;
 
-import com.nexters.goalpanzi.application.auth.dto.*;
+import com.nexters.goalpanzi.application.auth.dto.request.AppleLoginCommand;
+import com.nexters.goalpanzi.application.auth.dto.request.GoogleLoginCommand;
+import com.nexters.goalpanzi.application.auth.dto.request.RefreshTokenCommand;
+import com.nexters.goalpanzi.application.auth.dto.response.LoginResponse;
+import com.nexters.goalpanzi.application.auth.dto.response.TokenResponse;
 import com.nexters.goalpanzi.common.argumentresolver.LoginMemberId;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
