+++ src/main/java/com/nexters/goalpanzi/presentation/mission/MissionControllerDocs.java
@@ -1,6 +1,6 @@
 package com.nexters.goalpanzi.presentation.mission;
 
-import com.nexters.goalpanzi.application.mission.dto.MissionDetailResponse;
+import com.nexters.goalpanzi.application.mission.dto.response.MissionDetailResponse;
 import com.nexters.goalpanzi.common.argumentresolver.LoginMemberId;
 import com.nexters.goalpanzi.presentation.mission.dto.CreateMissionRequest;
 import io.swagger.v3.oas.annotations.Operation;
