+++ src/main/java/com/nexters/goalpanzi/presentation/mission/MissionVerificationControllerDocs.java
@@ -1,8 +1,7 @@
 package com.nexters.goalpanzi.presentation.mission;
 
-import com.nexters.goalpanzi.application.mission.dto.MissionResponse;
-import com.nexters.goalpanzi.application.mission.dto.MissionVerificationResponse;
-import com.nexters.goalpanzi.application.mission.dto.MissionsResponse;
+import com.nexters.goalpanzi.application.mission.dto.response.MissionVerificationResponse;
+import com.nexters.goalpanzi.application.mission.dto.response.MissionsResponse;
 import com.nexters.goalpanzi.common.argumentresolver.LoginMemberId;
 import com.nexters.goalpanzi.presentation.mission.dto.CreateMissionVerificationRequest;
 import io.swagger.v3.oas.annotations.Operation;
