+++ src/main/resources/application-dev.yml
@@ -1,11 +1,8 @@
 spring:
-  h2:
-    console:
-      enabled: true
-      path: /h2-console
   datasource:
-    url: jdbc:h2:~/test;MODE=MYSQL;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE;
-    username: sa
+    url: jdbc:mysql://localhost:3306/nexters
+    username: test
+    password: test
   jpa:
     database-platform: org.hibernate.dialect.MySQLDialect
     hibernate:
@@ -17,8 +14,9 @@ spring:
   flyway:
     enabled: true
     baseline-on-migrate: true
-    url: jdbc:h2:~/test;MODE=MYSQL;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE;
-    user: sa
+    url: jdbc:mysql://localhost:3306/nexters
+    user: test
+    password: test
     baseline-version: 0
 
 jwt:
