+++ src/test/java/com/nexters/goalpanzi/acceptance/LoginAcceptanceTest.java
@@ -3,9 +3,9 @@ package com.nexters.goalpanzi.acceptance;
 import com.nexters.goalpanzi.application.auth.SocialUserInfo;
 import com.nexters.goalpanzi.application.auth.SocialUserProvider;
 import com.nexters.goalpanzi.application.auth.SocialUserProviderFactory;
-import com.nexters.goalpanzi.application.auth.dto.AppleLoginCommand;
-import com.nexters.goalpanzi.application.auth.dto.GoogleLoginCommand;
-import com.nexters.goalpanzi.application.auth.dto.LoginResponse;
+import com.nexters.goalpanzi.application.auth.dto.request.AppleLoginCommand;
+import com.nexters.goalpanzi.application.auth.dto.request.GoogleLoginCommand;
+import com.nexters.goalpanzi.application.auth.dto.response.LoginResponse;
 import com.nexters.goalpanzi.fixture.TokenFixture;
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.Test;
